+++ sampleApps/express-complex/server.js
@@ -11,6 +11,7 @@ app.set('views', 'mvc/views') // set template dir
 app.set('view engine', 'html') // set teddy as default view engine
 if (!fs.existsSync('public')) fs.mkdirSync('public') // make the public folder if it does not exist
 app.use(express.static('public')) // make public folder serve static files
+fs.copyFileSync('styles.css', 'public/styles.css')
 
 // load shared express routes; this same file will be loaded in the frontend js bundle
 require('./mvc/routes')(app)
@@ -30,8 +31,10 @@ for (const file of fileList) {
   templates[file.slice(0, -5)] = templateCode // key is the file name with no file extension
 }
 fs.writeFileSync('public/templates.js', `module.exports = ${JSON.stringify(templates)}`)
-fs.writeFileSync('public/extra.css', `body { }`)
-fs.writeFileSync('public/extra.js', `console.log('extra script loaded!')`)
+
+// write extra files for testing certain routes
+fs.writeFileSync('public/extra.css', 'body { }')
+fs.writeFileSync('public/extra.js', 'console.log(\'extra script loaded!\')')
 
 // bundle frontend js
 const webpack = require('webpack')